Extract compute diagnostics BFF client from oversized bff.ts.
Keeps the facade from growing further past the 1k-line maintainability threshold for this feature.
